Move config.py to utils folder #272
