+++ open_mastr/__init__.py
@@ -1,4 +1,4 @@
-from open_mastr.soap_api.config import setup_project_home
+from open_mastr.utils.config import setup_project_home
 
 setup_project_home()
