+++ open_mastr/soap_api/sessions.py
@@ -18,7 +18,7 @@ __version__ = "v0.10.0"
 
 
 import os
-import open_mastr.soap_api.config as lc
+import open_mastr.utils.config as lc
 
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy import Column, Integer, String, Float, DateTime, Boolean
